+++ src/ModuleInstance/Contracts/Evaluator/ActivityInstanceEvaluator.php
@@ -18,18 +18,6 @@ interface ActivityInstanceEvaluator
 {
 
     /**
-     * Evaluate an activity instance for an administrator
-     *
-     * @param ActivityInstance $activityInstance Activity instance to evaluate
-     * @param User|null $user User to evaluate for
-     * @param Group|null $group Group to evaluate for
-     * @param Role|null $role Role to evaluate for
-     * 
-     * @return EvaluationContract[] Array of evaluations with the module instance id as the index
-     */
-    public function evaluateAdministrator(ActivityInstance $activityInstance, ?User $user = null, ?Group $group = null, ?Role $role = null);
-
-    /**
      * Evaluate an activity instance for a participant
      *
      * @param ActivityInstance $activityInstance Activity instance to evaluate