+++ src/Testing/ActivityInstance/LaravelAuthActivityInstanceResolver.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace BristolSU\Support\ActivityInstance;
+namespace BristolSU\Support\Testing\ActivityInstance;
 
 use BristolSU\Support\ActivityInstance\Contracts\ActivityInstanceResolver;
 use BristolSU\Support\ActivityInstance\Exceptions\NotInActivityInstanceException;
+use BristolSU\Support\ActivityInstance\ActivityInstance;
 use Illuminate\Contracts\Auth\Factory as AuthFactory;
 
 /**
@@ -13,16 +14,16 @@ class LaravelAuthActivityInstanceResolver implements ActivityInstanceResolver
 {
     /**
      * Laravel Auth Factory
-     * 
+     *
      * Used for saving and retrieving activity instances.
-     * 
+     *
      * @var AuthFactory
      */
     private $auth;
 
     /**
      * Initialise the activity instance resolver
-     * 
+     *
      * @param AuthFactory $auth
      */
     public function __construct(AuthFactory $auth)
@@ -32,7 +33,7 @@ class LaravelAuthActivityInstanceResolver implements ActivityInstanceResolver
 
     /**
      * Set the activity instance
-     * 
+     *
      * @param ActivityInstance $activityInstance Activity instance to set.
      */
     public function setActivityInstance(ActivityInstance $activityInstance)
@@ -41,11 +42,11 @@ class LaravelAuthActivityInstanceResolver implements ActivityInstanceResolver
     }
 
     /**
-     * 
+     *
      * Get the activity instance
-     * 
+     *
      * @return ActivityInstance Activity instance set through setActivityInstance
-     * 
+     *
      * @throws NotInActivityInstanceException If the activity instance is not set.
      */
     public function getActivityInstance(): ActivityInstance